simplify submodule
jj doesn't support submodules, so this is a more convenient approach.
Git submodules also didn't even make complete sense for this use case,
because there was no reason for the "child" repo to be pinned at a
specific revision in the parent repo. So now there is no need anymore
to create useless commits which bump the revision of the child repo.